// types.ts
export type FoodCategory = 'proteins' | 'carbohydrates' | 'vegetables' | 'fruits' | 'dairy' | 'snacks' | 'beverages';

interface LangString {
    en: string;
    ar: string;
}

export interface FoodItem {
    id: number;
    name: LangString;
    category: FoodCategory;
    description: LangString;
    nutrition: {
        calories: number;
        protein: number;
        carbs: number;
        fat: number;
    };
    details: {
        benefits: LangString;
        recommendation: LangString;
        allergens: LangString;
    };
}